refactor: project avatar and community
